Rename alert-table.tsx => AlertTable.tsx
