+++ extensions/ql-vscode/src/view/results/AlertTable.tsx
